% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_shannon.R
\name{my_shannon}
\alias{my_shannon}
\title{Shannon diversity}
\usage{
my_shannon(x)
}
\arguments{
\item{x}{A vector with values of abundance in the community}
}
\value{
Gives you a numeric vector with your intending shannon diversity
}
\description{
Shannon diversity
}
\examples{
num <- c(5, 8, 2)
my_shannon(num)
}
